blktap2: remove warnings.
This patch removes the following warnings on ia64.

> linux-2.6.18-xen.hg/drivers/xen/blktap2/device.c: In function
  'blktap_device_finish_request':
> linux-2.6.18-xen.hg/drivers/xen/blktap2/device.c:403: warning:
  format '%lld' expects type 'long long int', but argument 7 has type 'uint64_t'
> linux-2.6.18-xen.hg/drivers/xen/blktap2/sysfs.c: In function
  'blktap_sysfs_debug_device':
> linux-2.6.18-xen.hg/drivers/xen/blktap2/sysfs.c:276: warning: format
  '%llu' expects type 'long long unsigned int', but argument 4 has type
  'uint64_t'

+++ drivers/xen/blktap/device.c
@@ -400,7 +400,8 @@ blktap_device_finish_request(struct blktap *tap,
 	ret = res->status == BLKIF_RSP_OKAY ? 0 : -EIO;
 
 	BTDBG("req %p res status %d operation %d/%d id %lld\n", req,
-		res->status, res->operation, request->operation, res->id);
+	      res->status, res->operation, request->operation,
+	      (unsigned long long)res->id);
 
 	switch (request->operation) {
 	case BLKIF_OP_READ:
